pyproject.toml
[tool.poetry]
name = "s3-bucket-operator"
version = "0.1.0"
description = ""
authors = ["mingo gao <gxmk9999@gmail.com>"]
readme = "README.md"
packages = [{include = "s3_bucket_operator"}]

[tool.poetry.dependencies]
python = "^3.10"
aws-cdk-lib = "^2.70.0"
constructs = "^10.1.289"
pydantic = "^1.10.7"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
